BE/api/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
Pillow==8.4.0
drf-yasg==1.20.0
djangorestframework-simplejwt==5.1.0
mysqlclient==2.1.0
django-cors-headers==3.11.0